tools/flyway/requirements.txt
GitPython==2.1.8
mysqlclient==1.3.12
PyMySQL==0.8.0
PyYAML==3.12
